frontend/requirements.txt
streamlit==1.36.0
numpy==1.23.5
pandas==2.2.2
plotly==5.20.0
requests==2.31.0
python-dotenv
posthog==3.5.0
bs4==0.0.2
lxml==5.2.2